type topic

+++ kafka-stream/src/main/scala/org/esgi/project/streaming/models/Likes.scala
@@ -3,8 +3,8 @@ package org.esgi.project.streaming.models
 import play.api.libs.json.{Json, OFormat}
 
 case class Likes(
-                              _id: Float,
-                              score: Float,
+                              _id: Long,
+                              score: Long,
                             )
